[refactor] move param classes to folder
