+++ Backend/Interfaces/IAdminService.cs
@@ -1,5 +1,6 @@
 using System.Security.Claims;
 using Backend.Common;
+using Backend.Common.Params;
 using Backend.DTOs;
 
 namespace Backend.Interfaces;
