+++ Backend/Services/UserService.cs
@@ -1,4 +1,5 @@
 using Backend.Common;
+using Backend.Common.Params;
 using Backend.DTOs;
 using Backend.Extensions;
 using Backend.Interfaces;
